Created unittests for '/agents' API route
